+++ CHANGELOG.md
@@ -9,11 +9,13 @@
 
 - enhancements
   + [#45625](https://issues.afinedevelopment.com/issues/45625): add export to csv for index pages
+  + [#44718](https://issues.afinedevelopment.com/issues/44718): added counter for fields with a max length.
   + [#246508](https://issues.afinedevelopment.com/issues/246508): markdown helper text content updates
   + [#41106](https://issues.afinedevelopment.com/issues/41106): set max image and file upload
 
 - bugs
   + [#46027](https://issues.afinedevelopment.com/issues/46027]): fixed validation date picker bug
+  + [#46247](https://issues.afinedevelopment.com/issues/46247): remove periods from slugs
 
 
 ## 1.0.3